Fixing this error: "Cannot find module '@typescript-eslint/typescript-estree'" by moving from devDependencies to  dependencies (in package.json).
Why:
VibeSafe requires this package at runtime for scanning, but it was previously only installed for development. This caused errors when running the CLI as a global package.

No new dependencies produced, just a moving it so that it worked. When I tried to run 'vibesafe scan' on my machine, I encountered this error.